fix return type

+++ grails-app/services/com/unifina/service/SerializationService.groovy
@@ -43,7 +43,7 @@ class SerializationService {
 		return v
 	}
 
-	int serializationMaxBytes() {
+	long serializationMaxBytes() {
 		Long v = MapTraversal.getLong(grailsApplication.getConfig(), MAX_BYTES_CONFIG_KEY)
 		Assert.notNull(v, "Missing key \"" + MAX_BYTES_CONFIG_KEY + "\" from grailsApplication configuration")
 		return v
